Address navigation polish review feedback
